Adds some useful inventory functions to CivModCore's APIs.
The general use case for these functions are those that I discovered while working on ItemExchange and OldEnchanting; that if you only need a particular type of inventory, you can end up with a lot of repeating code while checking for nulls and inventory types. Say you're creating a plugin that modifies how Enchanting Tables work. How many events is that? How many times would you need to check that you're actually working with an enchanting inventory? How many copy pasted lines of code is that? This should hopefully reduce that to just a function call, because if it's not what you want it to be, it'll return null.